Improve gas estimation
