+++ lib/discord/alert.js
@@ -7,7 +7,7 @@ const CHANNELS_DEV = {
 };
 const CHANNELS_PROD = {
   action: "1154870152844820551",
-  warning: "1154870152844820551",
+  warning: "1156322646779248650",
 };
 
 const botToken = process.env.DISCORD_BOT_TOKEN;
