+++ lib/web3/gas/gasConfig.js
@@ -1,5 +1,5 @@
-import { parseUnits } from '@ethersproject/units'
-import { AVAILABLE_NETWORKS } from '../common'
+import { parseUnits } from "@ethersproject/units";
+import { AVAILABLE_NETWORKS } from "../common";
 
 // NOTE: Gas settings for transactions "maxFeePerGas" and "maxPriorityFeePerGas" will be used in this order:
 // 1. API call to explorer / gas station
@@ -9,46 +9,46 @@ import { AVAILABLE_NETWORKS } from '../common'
 const GAS_SETTINGS_BY_NETWORK = {
   // MAINNET:{},
   POLYGON: {
-    INITIAL_DROP_AMOUNT: '0.8',
-    DEFAULT_DROP_AMOUNT: '0.8',
-    MINIMUM_GAS_DROP_THRESHOLD: '0.3',
-    maxFeePerGas: parseUnits('190', 'gwei'),
-    maxPriorityFeePerGas: parseUnits('40', 'gwei'),
+    INITIAL_DROP_AMOUNT: "0.8",
+    DEFAULT_DROP_AMOUNT: "0.8",
+    MINIMUM_GAS_DROP_THRESHOLD: "0.3",
+    maxFeePerGas: parseUnits("190", "gwei"),
+    maxPriorityFeePerGas: parseUnits("40", "gwei"),
   },
   // MUMBAI:{GAS_TRACKER_URL:""},
   // GNOSIS:{},
   // GOERLI:{},
   OPTIMISM: {
-    INITIAL_DROP_AMOUNT: '0.00035',
-    DEFAULT_DROP_AMOUNT: '0.00065',
-    MINIMUM_GAS_DROP_THRESHOLD: '0.0003',
+    INITIAL_DROP_AMOUNT: "0.00035",
+    DEFAULT_DROP_AMOUNT: "0.00065",
+    MINIMUM_GAS_DROP_THRESHOLD: "0.0003",
   },
-  'OPTIMISM_GOERLI': {
-    INITIAL_DROP_AMOUNT: '0.00035',
-    DEFAULT_DROP_AMOUNT: '0.00065',
-    MINIMUM_GAS_DROP_THRESHOLD: '0.0003',
+  OPTIMISM_GOERLI: {
+    INITIAL_DROP_AMOUNT: "0.02",
+    DEFAULT_DROP_AMOUNT: "0.02",
+    MINIMUM_GAS_DROP_THRESHOLD: "0.005",
   },
   ARBITRUM: {
-    INITIAL_DROP_AMOUNT: '0.00035',
-    DEFAULT_DROP_AMOUNT: '0.00065',
-    MINIMUM_GAS_DROP_THRESHOLD: '0.0003',
+    INITIAL_DROP_AMOUNT: "0.00035",
+    DEFAULT_DROP_AMOUNT: "0.00065",
+    MINIMUM_GAS_DROP_THRESHOLD: "0.0003",
   },
-  'ARBITRUM_GOERLI': {
-    INITIAL_DROP_AMOUNT: '0.00035',
-    DEFAULT_DROP_AMOUNT: '0.00065',
-    MINIMUM_GAS_DROP_THRESHOLD: '0.0003',
+  ARBITRUM_GOERLI: {
+    INITIAL_DROP_AMOUNT: "0.00035",
+    DEFAULT_DROP_AMOUNT: "0.00065",
+    MINIMUM_GAS_DROP_THRESHOLD: "0.0003",
   },
   // SEPOLIA:{},
   // AVALANCHE:{},
   // AVALANCHE-FUJI:{},
   // MANTLE-WADSLEY:{},
-}
+};
 
 export const gasConfigByNetwork = ({ network }) => {
-  let config = GAS_SETTINGS_BY_NETWORK[network]
-  if (!config) config = GAS_SETTINGS_BY_NETWORK.POLYGON
-  return config
-}
+  let config = GAS_SETTINGS_BY_NETWORK[network];
+  if (!config) config = GAS_SETTINGS_BY_NETWORK.POLYGON;
+  return config;
+};
 
 const showMissingGasConfigWarning = () => {
   AVAILABLE_NETWORKS.map((networkName) => {
@@ -56,8 +56,8 @@ const showMissingGasConfigWarning = () => {
       /* eslint-disable-next-line no-console */
       console.log(
         `WARNING: Gas not configured for ${networkName}. Using default config for POLYGON.`
-      )
-  })
-}
+      );
+  });
+};
 
-showMissingGasConfigWarning()
+showMissingGasConfigWarning();
