+++ lib/web3/gas/index.js
@@ -103,117 +103,25 @@ export const checkGasAndAirdrop = async ({ address, amount, network }) => {
   }
 };
 
-export const getGasSettings = async ({ network }) => {
-  let alchemyNetwork;
-  try {
-    alchemyNetwork = getAlchemyNetworkEnum(network);
-  } catch (e) {
-    logger.debug(`getGasSettings() - no Alchemy network found for ${network}`);
-  }
-  let maxFeePerGas;
-  let maxPriorityFeePerGas;
+export const getGasSettings = async ({ unsignedTx, provider }) => {
+  // Estimate gas required
+  let gasUnits = await provider.estimateGas({ ...unsignedTx });
+  const gasUnitsPadded = gasUnits.mul(150).div(100); // add 50% for safety
 
-  // Use Alchemy first, if available for the network
-  if (alchemyNetwork) {
-    try {
-      const settings = {
-        apiKey: process.env.ALCHEMY_API_KEY,
-        network: getAlchemyNetworkEnum(network),
-      };
-      const alchemy = new Alchemy(settings);
-      let gasPrice = await alchemy.core.getGasPrice();
-      maxFeePerGas = gasPrice;
-      maxPriorityFeePerGas = gasPrice;
-    } catch (e) {
-      if (network === "POLYGON") {
-        maxFeePerGas = DEFAULT_GAS_MAX_PRIORITY_FEE_PER_GAS;
-        maxPriorityFeePerGas = DEFAULT_GAS_MAX_PRIORITY_FEE_PER_GAS;
-        logger.error(`Falling back to default gas settings for Polygon`);
-      }
-      logger.error(`Can't get gas settings from Alchemy for ${network}`);
-      logger.error(e);
-    }
+  // Current gas price
+  const gasPrice = await provider.getGasPrice();
+  const txFee = gasUnitsPadded.mul(gasPrice);
+
+  // Check gas balance
+  const signerBalance = await provider.getBalance();
+  if (signerBalance.lt(txFee)) {
+    throw Error(`Insufficient gas balance`);
   }
 
-  // Using Infura for Avalanche and Avalanche-Fuji
-  if (network === "AVALANCHE" || network === "AVALANCHE-FUJI") {
-    try {
-      if (network === "AVALANCHE") {
-        axios
-          .post(
-            `https://avalanche-mainnet.infura.io/v3/${process.env.INFURA_ID}`,
-            {
-              jsonrpc: "2.0",
-              id: 1,
-              method: "eth_gasPrice",
-              params: [],
-            },
-            {
-              headers: { "Content-Type": "application/json" },
-            }
-          )
-          .then(
-            (response) => {
-              maxFeePerGas = response.data.result;
-              maxPriorityFeePerGas = response.data.result;
-            },
-            (error) => {
-              logger.error(`Can't get gas settings from Infura for ${network}`);
-              logger.error(error);
-            }
-          );
-      } else {
-        axios
-          .post(
-            `https://avalanche-fuji.infura.io/v3/${process.env.INFURA_ID}`,
-            {
-              jsonrpc: "2.0",
-              id: 1,
-              method: "eth_gasPrice",
-              params: [],
-            },
-            {
-              headers: { "Content-Type": "application/json" },
-            }
-          )
-          .then(
-            (response) => {
-              maxFeePerGas = response.data.result;
-              maxPriorityFeePerGas = response.data.result;
-            },
-            (error) => {
-              logger.error(`Can't get gas settings from Infura for ${network}`);
-              logger.error(error);
-            }
-          );
-      }
-    } catch (e) {
-      logger.error(`Can't get gas settings from Infura for ${network}`);
-      logger.error(e);
-    }
-  }
-
-  // Gnosis
-  if (network === "GNOSIS") {
-    try {
-      let response = await axios.get(
-        "https://api.gnosisscan.io/api?module=proxy&action=eth_gasPrice"
-      );
-      let gasPrice = response.data.result;
-      maxFeePerGas = BigNumber.from(gasPrice);
-      maxPriorityFeePerGas = maxFeePerGas;
-    } catch (e) {
-      logger.error(`Can't get gas settings for Gnosis`);
-      logger.error(e);
-    }
-  }
-
-  if (!maxFeePerGas || !maxPriorityFeePerGas) {
-    maxFeePerGas = undefined;
-    maxPriorityFeePerGas = undefined;
-  }
-
-  return { maxFeePerGas, maxPriorityFeePerGas };
+  // Fee settings
+  const feeData = await provider.getFeeData();
+  const { maxFeePerGas, maxPriorityFeePerGas } = feeData;
+  return { maxFeePerGas, maxPriorityFeePerGas, gasLimit: gasUnits };
 };
 
 const notifyDiscord = async ({ address, txHash, payerBalance, network }) => {
