+++ lib/web3/oracle.js
@@ -79,6 +79,7 @@ export const updateResult = async (network, cType) => {
   const collateral = Object.values(geb.tokenList).find(
     (t) => t.bytes32String === cType
   );
+  let txResponse;
   try {
     const params = await geb.contracts.oracleRelayer.cParams(cType);
     const { oracle: delayedOracleAddress } = params;
@@ -95,7 +96,7 @@ export const updateResult = async (network, cType) => {
       network,
     });
 
-    const txResponse = await botSendTx({ unsigned: txData, network });
+    txResponse = await botSendTx({ unsigned: txData, network });
     await tx.update({ hash: txResponse.hash });
     await txResponse.wait();
     return txResponse.hash;
@@ -105,7 +106,15 @@ export const updateResult = async (network, cType) => {
       embed: {
         color: 15548997,
         title: `🦉 DelayedOracle 🚫 FAILED | ${network}`,
-        description: `updateResult() failed for ${collateral?.symbol} with error: ${e} `,
+        description: `updateResult() failed for ${collateral?.symbol}
+${
+  txResponse?.hash
+    ? `[receipt](${getExplorerBaseUrlFromName(network)}tx/${txResponse.hash})
+\`\`\`js
+${e}
+\`\`\``
+    : ""
+}`,
         footer: { text: new Date().toString() },
       },
       channelName: "warning",
@@ -120,6 +129,7 @@ export const updateCollateralPrice = async (network, cType) => {
   const collateral = Object.values(geb.tokenList).find(
     (t) => t.bytes32String === cType
   );
+  let txResponse;
   try {
     const txData =
       await geb.contracts.oracleRelayer.populateTransaction.updateCollateralPrice(
@@ -133,7 +143,7 @@ export const updateCollateralPrice = async (network, cType) => {
       network,
     });
 
-    const txResponse = await botSendTx({ unsigned: txData, network });
+    txResponse = await botSendTx({ unsigned: txData, network });
     await tx.update({ hash: txResponse.hash });
     await txResponse.wait();
     return txResponse.hash;
@@ -143,7 +153,17 @@ export const updateCollateralPrice = async (network, cType) => {
       embed: {
         color: 15548997,
         title: `🦉 OracleRelayer 🚫 FAILED | ${network}`,
-        description: `updateCollateralPrice() failed for collateral ${collateral?.symbol} with error: ${e} `,
+        description: `updateCollateralPrice() failed for collateral ${
+          collateral?.symbol
+        }
+${
+  txResponse?.hash
+    ? `[receipt](${getExplorerBaseUrlFromName(network)}tx/${txResponse.hash})
+\`\`\`js
+${e}
+\`\`\``
+    : ""
+}`,
         footer: { text: new Date().toString() },
       },
       channelName: "warning",
