+++ lib/web3/wallets/bot.js
@@ -28,22 +28,14 @@ export const botSendTx = async ({ unsigned, network }) => {
   try {
     await checkGasAndAirdrop({ network, address: botWallet.address, network });
     const botWalletWithProvider = botWallet.connect(Web3Providers[network]);
-    const gasSettings = await getGasSettings({ network });
-
-    let maxFeePerGas = gasSettings.maxFeePerGas;
-    let maxPriorityFeePerGas = gasSettings.maxPriorityFeePerGas;
-
-    if (network === "ARBITRUM_GOERLI") {
-      // Arbitrum Goerli gets a 10x gas multiplier, due to failure to estimate
-      // updateCollateralPrice() gas usage
-      maxFeePerGas = maxFeePerGas.mul(200);
-      maxPriorityFeePerGas = maxPriorityFeePerGas.mul(200);
-    }
+    const gasSettings = await getGasSettings({
+      unsignedTx: unsigned,
+      provider: botWalletWithProvider,
+    });
 
     return await botWalletWithProvider.sendTransaction({
       ...unsigned,
-      maxFeePerGas,
-      maxPriorityFeePerGas,
+      ...gasSettings,
     });
     return null;
   } catch (e) {
