+++ pages/api/rate.js
@@ -1,5 +1,4 @@
-import { updateRate } from "../../lib/web3/geb";
-import { updateFeed } from "../../lib/web3/oracle";
+import { updateRate } from "../../lib/web3/rate";
 
 const ARBITRUM_GOERLI = "ARBITRUM_GOERLI";
 
@@ -12,9 +11,6 @@ export default async function handler(request, response) {
   let network = ARBITRUM_GOERLI;
   if (request.query.network) network = request.query.network;
   await updateRate(network);
-
-  // TODO: also update oracle feed
-  // await updateFeed(OPTIMISM_GOERLI);
 
   response.status(200).json({ success: true });
 }
